Tests/LibWeb/Text/expected/setTimeout-string-handler.txt

string timer code point: 233
